fix: evento do botao nativo de voltar
