+++ lib/core/routes.dart
@@ -3,5 +3,4 @@ class Routes {
   static const String start = '/start';
   static const String home = '/home';
   static const String login = '/login';
-  static const String tellGame = '/tell-game';
 }
