+++ lib/pages/learning_screen/components/story_card.dart
